[add] 홈 스크린 파일을 생성합니다.
